Add files via upload
JS30, D6: Place Population

Data is fetched, the response is then parsed, pushed, and spread into an array

This is a project to revisit because I ran into issues using arrow functions while trying to extract the value.  The project video used regular functions but I'd like to understand functions and scope at a deeper level.